refactor: make providers simpler and lighter
