+++ packages/core/src/plugins/basic.ts
@@ -63,6 +63,12 @@ export const pluginBasic = (): RsbuildPlugin => ({
               path.resolve(info.absoluteResourcePath).replace(/\\/g, '/'),
           );
         }
+
+        process.env.RSPACK_CONFIG_VALIDATE = 'loose-silent';
+
+        // improve kill process performance
+        // https://github.com/web-infra-dev/rspack/pull/5486
+        process.env.WATCHPACK_WATCHER_LIMIT ||= '20';
       },
     );
   },
