+++ packages/shared/src/chain.ts
@@ -10,7 +10,6 @@ import {
 } from './constants';
 import { getDistPath, getFilename } from './fs';
 import { debug } from './logger';
-import { mergeChainedOptions } from './mergeChainedOptions';
 import type {
   BundlerChain,
   BundlerChainRule,
@@ -22,11 +21,10 @@ import type {
   RsbuildContext,
   RsbuildEntry,
   RsbuildPluginAPI,
-  RsbuildTarget,
   RspackConfig,
 } from './types';
 import { addTrailingSlash, isPlainObject, removeTailingSlash } from './utils';
-import { castArray, ensureAbsolutePath } from './utils';
+import { castArray } from './utils';
 
 export async function getBundlerChain() {
   const { default: WebpackChain } = await import(
@@ -390,103 +388,6 @@ export function applyOutputPlugin(api: RsbuildPluginAPI) {
       }
     },
   );
-}
-
-export function applyResolvePlugin(api: RsbuildPluginAPI) {
-  api.modifyBundlerChain({
-    order: 'pre',
-    handler: (chain, { target, CHAIN_ID }) => {
-      const config = api.getNormalizedConfig();
-
-      applyExtensions({ chain });
-
-      applyAlias({
-        chain,
-        target,
-        config,
-        rootPath: api.context.rootPath,
-      });
-
-      // in some cases (modern.js), get error when fullySpecified rule after js rule
-      applyFullySpecified({ chain, config, CHAIN_ID });
-    },
-  });
-}
-
-// compatible with legacy packages with type="module"
-// https://github.com/webpack/webpack/issues/11467
-function applyFullySpecified({
-  chain,
-  CHAIN_ID,
-}: {
-  chain: BundlerChain;
-  config: NormalizedConfig;
-  CHAIN_ID: ChainIdentifier;
-}) {
-  chain.module
-    .rule(CHAIN_ID.RULE.MJS)
-    .test(/\.m?js/)
-    .resolve.set('fullySpecified', false);
-}
-
-function applyExtensions({ chain }: { chain: BundlerChain }) {
-  const extensions = [
-    // most projects are using TypeScript, resolve .ts(x) files first to reduce resolve time.
-    '.ts',
-    '.tsx',
-    '.js',
-    '.jsx',
-    '.mjs',
-    '.json',
-  ];
-
-  chain.resolve.extensions.merge(extensions);
-}
-
-function applyAlias({
-  chain,
-  target,
-  config,
-  rootPath,
-}: {
-  chain: BundlerChain;
-  target: RsbuildTarget;
-  config: NormalizedConfig;
-  rootPath: string;
-}) {
-  const { alias } = config.source;
-
-  if (!alias) {
-    return;
-  }
-
-  const mergedAlias = mergeChainedOptions({
-    defaults: {},
-    options: alias,
-    utils: { target },
-  });
-
-  /**
-   * Format alias value:
-   * - Relative paths need to be turned into absolute paths.
-   * - Absolute paths or a package name are not processed.
-   */
-  for (const name of Object.keys(mergedAlias)) {
-    const values = castArray(mergedAlias[name]);
-    const formattedValues = values.map((value) => {
-      if (typeof value === 'string' && value.startsWith('.')) {
-        return ensureAbsolutePath(rootPath, value);
-      }
-      return value;
-    });
-
-    chain.resolve.alias.set(
-      name,
-      (formattedValues.length === 1 ? formattedValues[0] : formattedValues) as
-        | string
-        | string[],
-    );
-  }
 }
 
 export function chainToConfig(chain: BundlerChain): RspackConfig {
